ArrayList.cs
Program which can produce numbers between two given numbers. Explanation below:

I have two numbers, suppose 100 and 200, in the positions 1  and 2 respectively, need an algorithm which will take "position" as an input and produce a number that can fit in between the numbers 

So, if I give 2 as "position", then the algorithm should produce a number which is greater than 100 and less than 200. Suppose we got 150, it will occupy the position 2 and 200 will shift to position 3